Revert "Export an abstraction instead of the implementations"
