+++ lib/src/tsid_stub.dart
@@ -1,23 +1,30 @@
 import 'dart:typed_data';
 
-interface class Tsid {
-  Tsid(final BigInt number) {
+class Tsid {
+  static Uint8List initializeAlphabetValues() {
     throw Exception("Stub implementation");
   }
 
-  Tsid.fromNumber(final BigInt number);
-
-  Tsid.fromBytes(Uint8List bytes);
-
-  Tsid.fromString(String string);
-
-  BigInt toLong() {
+  static int getNumberFromBytes(Uint8List bytes) {
     throw Exception("Stub implementation");
   }
 
-  int toInt() {
+  static int getNumberFromString(String string) {
     throw Exception("Stub implementation");
   }
+
+  Tsid(final int number) {
+    throw Exception("Stub implementation");
+  }
+
+  Tsid.fromNumber(final int number) : this(number);
+
+  Tsid.fromBytes(Uint8List bytes) : this(getNumberFromBytes(bytes));
+
+  Tsid.fromString(String string) : this(getNumberFromString(string));
+
+  int toLong() {
+    throw Exception("Stub implementation");  }
 
   Uint8List toBytes() {
     throw Exception("Stub implementation");
@@ -29,24 +36,19 @@ interface class Tsid {
 
   @override
   String toString() {
-    throw Exception("Stub implementation");
-  }
+    throw Exception("Stub implementation");  }
 
   String toLowerCase() {
-    throw Exception("Stub implementation");
-  }
+    throw Exception("Stub implementation");  }
 
-  BigInt getUnixMilliseconds(final BigInt customEpoch) {
-    throw Exception("Stub implementation");
-  }
+  int getUnixMilliseconds(final int customEpoch) {
+    throw Exception("Stub implementation");  }
 
-  BigInt getTime() {
-    throw Exception("Stub implementation");
-  }
+  int getTime() {
+    throw Exception("Stub implementation");  }
 
-  BigInt getRandom() {
-    throw Exception("Stub implementation");
-  }
+  int getRandom() {
+    throw Exception("Stub implementation");  }
 
   static bool isValid(final String string) {
     throw Exception("Stub implementation");
@@ -83,6 +85,7 @@ interface class Tsid {
     throw Exception("Stub implementation");
   }
 
+
   factory Tsid.getTsid() {
     throw Exception("Stub implementation");
   }
@@ -106,19 +109,5 @@ interface class Tsid {
       return false;
     }
     return super.hashCode == other.hashCode;
-  }
-}
-
-interface class TsidFactory {
-  factory TsidFactory() {
-    throw Exception("Stub implementation");
-  }
-
-  factory TsidFactory.fromNode(BigInt node) {
-    throw Exception("Stub implementation");
-  }
-
-  Tsid create() {
-    throw Exception("Stub implementation");
   }
 }
